move assign algs + add check for mixed hypotheses
